requirements.txt
altgraph==0.17.3
click==8.1.3
pkg_resources==0.0.0
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.15
python-dotenv==0.21.0
